refactor: kill dead code. refactor: improve file hierarchy and imports
