+++ browser/UI/strike.js
