+++ browser/assets/soundEffects.js
