+++ browser/index.js
@@ -14,8 +14,8 @@ import introText from './introText.js'
 import {mediaRecorder, startRecording, stopRecording} from './audio/audio'
 import loadAllAssets from './assets/assets'
 import { startSyncingPhaseAndStrikes } from './firebase'
-import store from './store.jsx'
-import { setNavigatorStatus, setDriverStatus } from './reducers/strike-phase.js'
+import store from './store'
+import { setNavigatorStatus, setDriverStatus } from './reducers/strike-phase'
 import 'aframe-daydream-controller-component'
 import stopDefaultAndPropagation from './utils/events'
 import setUpDayDreamAudio from './utils/headset'
