+++ browser/store.js
@@ -1,6 +1,6 @@
 import { createStore, applyMiddleware } from 'redux'
 import { composeWithDevTools } from 'redux-devtools-extension'
-import reducer from './reducers/strike-phase.js'
+import reducer from './reducers/strike-phase'
 import {createLogger} from 'redux-logger'
 import thunkMiddleware from 'redux-thunk'
